systems/perception/metrics/requirements.txt
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
narwhals==1.43.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.0
plotly==6.1.2
polling2==0.5.0
protobuf==6.31.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
resim-open-core==0.14.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
